+++ gfx/thebes/gfxFont.h
@@ -133,8 +133,8 @@ struct gfxFontStyle {
     // Not meant to be called when sizeAdjust = 0.
     gfxFloat GetAdjustedSize(gfxFloat aspect) const {
         NS_ASSERTION(sizeAdjust != 0.0, "Not meant to be called when sizeAdjust = 0");
-        gfxFloat adjustedSize = std::max(NS_round(size*(sizeAdjust/aspect)), 1.0);
-        return std::min(adjustedSize, FONT_MAX_SIZE);
+        gfxFloat adjustedSize = std::max(NS_round(size*(sizeAdjust/aspect)), (double)1.0);
+        return std::min(adjustedSize, (double)FONT_MAX_SIZE);
     }
 
     PLDHashNumber Hash() const {
