+++ js/src/jsstr.cpp
@@ -1236,7 +1236,7 @@ str_contains(JSContext *cx, unsigned argc, Value *vp)
             double d;
             if (!ToInteger(cx, args[1], &d))
                 return false;
-            pos = uint32_t(Min(Max(d, 0.0), double(UINT32_MAX)));
+            pos = uint32_t(Min(Max(d, (double)0.0), double(UINT32_MAX)));
         }
     }
 
@@ -1287,7 +1287,7 @@ str_indexOf(JSContext *cx, unsigned argc, Value *vp)
             double d;
             if (!ToInteger(cx, args[1], &d))
                 return false;
-            pos = uint32_t(Min(Max(d, 0.0), double(UINT32_MAX)));
+            pos = uint32_t(Min(Max(d, (double)0.0), double(UINT32_MAX)));
         }
     }
 
@@ -1415,7 +1415,7 @@ str_startsWith(JSContext *cx, unsigned argc, Value *vp)
             double d;
             if (!ToInteger(cx, args[1], &d))
                 return false;
-            pos = uint32_t(Min(Max(d, 0.0), double(UINT32_MAX)));
+            pos = uint32_t(Min(Max(d, (double)0.0), double(UINT32_MAX)));
         }
     }
 
@@ -1472,7 +1472,7 @@ str_endsWith(JSContext *cx, unsigned argc, Value *vp)
             double d;
             if (!ToInteger(cx, args[1], &d))
                 return false;
-            pos = uint32_t(Min(Max(d, 0.0), double(UINT32_MAX)));
+            pos = uint32_t(Min(Max(d, (double)0.0), double(UINT32_MAX)));
         }
     }
